refactor: move value parser plugin to own file
